% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.R
\name{gen.imp.resp}
\alias{gen.imp.resp}
\title{Obtain imputed continuous responses from ordinal responses}
\usage{
gen.imp.resp(data, num.iter = 5)
}
\arguments{
\item{data}{a data.frame containing ordinal variables (e.g. survey responses).}

\item{num.iter}{a number specifying the number of times to iterate the imputation (defaults to 20)}
}
\value{
A data.frame containing the imputed responses.
}
\description{
Given a dataset of ordinal values (e.g. survey responses), obtain continuous imputed responses based
on quantiles of N(0,1), and using all other columns (similar to the MICE package).
}
\examples{
imputed <- gen.imp.resp(multiis, num.iter = 1)

head(imputed)
}
