% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.R
\name{gen.latent.datasets}
\alias{gen.latent.datasets}
\title{Generate multiple datasets of latent underlying variables}
\usage{
gen.latent.datasets(M, data, grp.indicator, scores = "Bartlett",
  num.iter = 5)
}
\arguments{
\item{M}{number of datasets to construct.}

\item{data}{a (non-empty) numeric vector of data values.}

\item{grp.indicator}{a (non-empty) numeric vector of data values.}

\item{scores}{type of score to use.}

\item{num.iter}{number of iterations to use in imputation step}
}
\value{
A list containing datasets with latent underlying variables
}
\description{
Given ordinal responses to survey questions, we generate multiple
underlying latent variable datasets
}
\examples{
# Create indicators (a label indicating which latent variable the question corresponds to)
grp.indicator <- sapply(names(multiis), FUN =
                         function(x){strsplit(x, split = "_")[[1]][2]})

latent.datasets <- gen.latent.datasets(5, multiis, grp.indicator = grp.indicator, num.iter = 1)

head(latent.datasets[[1]])

head(latent.datasets[[2]])
}
