% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.R
\name{gen.latent.vars}
\alias{gen.latent.vars}
\title{Generate latent underlying variables given latent question responses}
\usage{
gen.latent.vars(data, grp.indicator, scores = "Bartlett")
}
\arguments{
\item{data}{a (non-empty) numeric vector of data values.}

\item{grp.indicator}{a (non-empty) numeric vector of data values.}

\item{scores}{number of bootstrap resamples to perform.}
}
\value{
A data.frame containing the underlying latent variable estimates.
}
\description{
Given imputed latent variable responses to survey questions, we generate the
underlying latent variables that correspond to those questions
}
\examples{
# Create indicators (a label indicating which latent variable the question corresponds to)
grp.indicator <- sapply(names(imputed), FUN =
                         function(x){strsplit(x, split = "_")[[1]][2]})

latent.vars <- gen.latent.vars(imputed, grp.indicator = grp.indicator)

head(latent.vars)
}
