% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiis-data.R
\docType{data}
\name{multiis}
\alias{multiis}
\title{Data of responses to The Multicultural Identity Integration Scale (MULTIIS)}
\format{A data.frame}
\usage{
data(multiis)
}
\description{
Data from the survey described in The Multicultural Identity Integration
Scale (MULTIIS): Developing a Comprehensive Measure for Configuring One’s Multiple
Cultural Identities Within the Self
}
\examples{
data(multiis)

head(multiis)
}
\references{
Yampolsky et al. (2016) Cultural Diversity and Ethnic Minority Psychology http://dx.doi.org/10.1037/cdp0000043
}
\keyword{datasets}
