% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.R
\name{pool.analyses}
\alias{pool.analyses}
\title{Pool analyses results given M latent variable data sets, and estimate parameters}
\usage{
pool.analyses(latent.datasets)
}
\arguments{
\item{latent.datasets}{a (non-empty) list of lists returned by the gen.latent.vars function.}
}
\value{
A list of lists of 5 elements consisting of: within-imputation variance, between-imputation variance, total variance, number of analyses, and a vector of final estimated parameters
}
\description{
Pool analyses results given M latent variable data sets, and estimate parameters
}
\examples{

}
