% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.R
\name{pooled.cor.test}
\alias{pooled.cor.test}
\title{Perform a correlation test on multiply imputed datasets}
\usage{
pooled.cor.test(datasets, indices = c(1, 2), alternative = "two.sided",
  method = "pearson")
}
\arguments{
\item{datasets}{a list of data.frames}

\item{indices}{a vector of length two specifying the index of the columns on which to test correlation}

\item{alternative}{the alternative hypothesis}

\item{method}{the method to compute correlation. Currently only "pearson" is supported.}

\item{grp.indicator}{a vector indicating which underlying group the columns in each dataset correpsonds to}
}
\value{
A list containing the mean correlation and a p-value from the hypothesis test
}
\description{
Given imputed datasets, perform a correlation test between specified columns
}
\examples{
# Create indicators (a label indicating which latent variable the question corresponds to)
grp.indicator <- sapply(names(multiis), FUN =
                         function(x){strsplit(x, split = "_")[[1]][2]})

latent.datasets <- gen.latent.datasets(5, multiis, grp.indicator = grp.indicator, num.iter = 1)

pooled.cor.test(latent.datasets, indices = c(1,3))
}
