% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim.R
\name{gen.latent.vars}
\alias{gen.latent.vars}
\title{Generate latent underlying variables given imputed question responses}
\usage{
gen.latent.vars(data, grp.indicator, scores = "Bartlett", num.iter = 20)
}
\arguments{
\item{data}{a (non-empty) numeric vector of data values.}

\item{grp.indicator}{a (non-empty) numeric vector of data values.}

\item{scores}{type of score to use.}

\item{num.iter}{number of iterations to use in imputation step}
}
\value{
A list l of length 2, where l[[1]] is a data.frame containing the underlying
latent variable estimates, and l[[2]] is a vector of j factor analyses objects
(j is the total number of latent variables).
}
\description{
Given ordinal responses to survey questions, we generate the
underlying latent variables that correspond to those questions
}
\examples{
setwd("~/GitProjects/440proj") # Set to project folder
multiis <- read.csv("data/MULTIIS.csv")
imputed <- gen.imp.resp(multiis)
# Create indicators (a label indicating which latent variable the question corresponds to)
grp.indicator <- sapply(names(imputed), FUN =
                         function(x){strsplit(x, split = "_")[[1]][2]})

latent.vars <- gen.latent.vars(imputed, grp.indicator = grp.indicator)

head(latent.vars)
}
